+++ src/components/recipeCreate/element/RecipeIngredientColorList.jsx
@@ -1,86 +1,41 @@
-import React from "react";
-import Slider from "react-slick";
-import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
-
-import RecipeIngredientColor from "./RecipeIngredientColor";
-
 import styled from "styled-components";
 
-const RecipeIngredientColorList = (props) => {
-  const {colorLists, idx, register} = props
-
-  const settings = {
-    dots: true,
-    infinite: false,
-    speed: 500,
-    slidesToShow: 1,
-    slidesToScroll: 1,
-    draggable: true
-  };
-
+const RecipeIngredientColorList = (
+  { colorList, name, register, config={}, onChange=null, onClick=null }
+) => {
 	return (
-    <div className="container">
-      <StSlick>
-        <Slider {...settings}>
-          <StRecipeIngredientColorList>
-            {colorLists[0].map((color, i)=>{
-              return (
-                <RecipeIngredientColor
-                  key={i}
-                  color={color}
-                  name={`ingredientList.${idx}.ingredientColor`}
-                  register={register} 
-                />
-              )
-            })}
-          </StRecipeIngredientColorList>
-          <StRecipeIngredientColorList>
-            {colorLists[1].map((color, i)=>{
-              return (
-                <RecipeIngredientColor
-                  key={i}
-                  color={color}
-                  name={`ingredientList.${idx}.ingredientColor`}
-                  register={register} 
-                />
-              )
-            })}
-          </StRecipeIngredientColorList>
-          <StRecipeIngredientColorList>
-            {colorLists[2].map((color, i)=>{
-              return (
-                <RecipeIngredientColor
-                  key={i}
-                  color={color}
-                  name={`ingredientList.${idx}.ingredientColor`}
-                  register={register} 
-                />
-              )
-            })}
-          </StRecipeIngredientColorList>
-        </Slider>
-      </StSlick>
-      <div>
-      </div>
-    </div>
+    <StRecipeIngredientColorList>
+    {colorList.map((color, i)=>{
+      return(
+        <>
+          <input 
+            className="colorLabel"
+            id={ name+color }
+            type="radio" 
+            value={ color }
+            {...register( name, config )}
+            onChange={ onChange } 
+          />
+          
+          <StColorCircle 
+            className="colorLabel"
+            htmlFor={ name+color }
+            color={ color }
+            onClick={ onClick } 
+          />
+        </>
+      )
+    })}
+    </StRecipeIngredientColorList>
   )
 }
 
 export default RecipeIngredientColorList;
 
-const StSlick = styled.div`
-  width: 100%;
-  height: 80px;
-  
-  position: absolute;
-  left: 0;
-  `
-
 const StRecipeIngredientColorList = styled.div`
   width: 100%;
   padding-top: 3px;
-
+  
   input {
     position: absolute;
     /* opacity: 0;
@@ -94,5 +49,31 @@ const StRecipeIngredientColorList = styled.div`
   input:checked + .colorLabel {
     box-shadow: 0 2px 7px 3px rgba(45, 35, 53, 0.1);
     transform: translateY(-2px) scale(1.1);
+  }
+`
+
+const StColorCircle = styled.label`
+  &.colorLabel {
+    width: 30px;
+    height: 30px;
+    border-radius: 50%;
+    
+    margin: 0 5px;
+    box-sizing: initial;
+    
+    display: inline-block;
+    
+    background-color: ${props => props.color};
+    border: 4px solid #fff;
+    
+    cursor: pointer;
+  }
+  
+  @media (max-width: 500px) {
+    &.colorLabel {
+      width: 22px;
+      height: 22px;
+      border: 3px solid #fff;
+    }
   }
 `
